Change price and price_to columns to BIGINT in migrations

+++ db/migrations/000090_ad_pricing_flags.down.sql
@@ -1,6 +1,6 @@
 UPDATE ad SET price = 0 WHERE price IS NULL;
 ALTER TABLE ad
-    MODIFY price DECIMAL(10, 2) NOT NULL,
+    MODIFY price BIGINT NOT NULL,
     DROP COLUMN hide_phone,
     DROP COLUMN negotiable,
     DROP COLUMN price_to,
